Remove bad check in cronbach_alpha calculation
There is no reason to return nil if a single vector has a 0 variance.  For example, let's say you are giving a test, and every single taker gets the easiest question correct.  The variance for that question vector is 0, but Cronbach's alpha can still be calculated correctly for the entire dataset.